requirements.txt
# Dependências do 3D Point Cloud Viewer

# OpenGL e GLFW para renderização
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
glfw>=2.0.0

# Computação numérica
numpy>=1.20.0

# JIT compilation para performance
numba>=0.50.0
